requirements.txt

moviepy>=1.0.3
pytube>=15.0.0
Pillow>=10.0.0
requests>=2.31.0
pysocks>=1.7.1
numpy>=1.24.0